+++ src/baobab_scryfall_api_caller/models/bulk_data/__init__.py
@@ -1,5 +1,6 @@
 """Modeles du domaine Bulk Data."""
 
 from baobab_scryfall_api_caller.models.bulk_data.bulk_data import BulkData
+from baobab_scryfall_api_caller.models.bulk_data.bulk_download_result import BulkDownloadResult
 
-__all__ = ["BulkData"]
+__all__ = ["BulkData", "BulkDownloadResult"]