+++ src/baobab_scryfall_api_caller/services/bulk_data/__init__.py
@@ -2,5 +2,8 @@
 
 from baobab_scryfall_api_caller.services.bulk_data.bulk_data_api_client import BulkDataApiClient
 from baobab_scryfall_api_caller.services.bulk_data.bulk_data_service import BulkDataService
+from baobab_scryfall_api_caller.services.bulk_data.bulk_dataset_downloader import (
+    BulkDatasetDownloader,
+)
 
-__all__ = ["BulkDataApiClient", "BulkDataService"]
+__all__ = ["BulkDataApiClient", "BulkDataService", "BulkDatasetDownloader"]